Optimization, correction and compatibility.
Refactorized the whole functions. Tested with (setq use-empty-active-region t) and works fine with or without it. Optimized, tested and made absolutely sure that it executes ov-clear and ov-set strictly if the region changes given any command secuence you can think of. Note: my previous commit was pretty bad, but this one is a definitive one and have the two basics that I expected: good performance, a little delay in case of very fast repetitions so that it wont perform highlightings that the user wouldn't even notice. Also the code is as clear and simple as I could make it, for future extensibility.